CLN: conform read_clipboard / to_clipboard to new io standards
     removed to io.clipboard (from io.parsers)
